feat: Update message status display in chat.html
This commit adds functionality to update the display of message status in the chat.html file of the SkillSetu Chat application. When a receipt update is received through the WebSocket connection, the corresponding message element is updated with the appropriate status icon. This improvement enhances the user experience by providing real-time updates on message statuses.