refactor: Improve WebSocket message handling and receipt updates
The code changes in  and  files improve the handling of WebSocket messages and receipt updates. It introduces a new message type for receipt updates and adds logic to send receipt updates to the  when a message is received or read. The changes also include updating the  method in the  to accept an additional  parameter. This allows for stopping the propagation of receipt updates when a message is read.